temporary fix data local only for guest

+++ src/features/WatchlistSection/api/deleteWatchlistItem.ts
@@ -5,6 +5,13 @@ import { createClient } from "@/utils/supabase/server";
 export async function deleteWatchlistItem(id: number) {
     const supabase = await createClient();
 
+    const { data: authData, error: authError } = await supabase.auth.getUser();
+    if (authError) throw authError;
+    if (!authData.user) throw new Error("Not logged in");
+    if (authData.user.email === "guest@quickflick.com") {
+        return "guest";
+    }
+
     const { data, error } = await supabase.from("watchlist").delete().eq("id", id);
 
     if (error) {
